backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.2.0
pdf2image>=1.16.3
reportlab>=4.0.7
wikipedia-api>=0.6.0
openai>=1.3.5
python-dotenv>=1.0.0
transformers>=4.35.2
torch>=2.1.1
youtube-transcript-api>=0.6.1
requests>=2.28.0
